Isotopic envelop now represented by double instead of float values.

+++ InformedProteomics.Backend/Data/Spectrometry/Isotope.cs
@@ -4,14 +4,14 @@ namespace InformedProteomics.Backend.Data.Spectrometry
 {
     public class Isotope: IComparable<Isotope>
     {
-        public Isotope(int index, float ratio)
+        public Isotope(int index, double ratio)
         {
             Index = index;
             Ratio = ratio;
         }
 
         public int Index { get; private set; }
-        public float Ratio { get; private set; }
+        public double Ratio { get; private set; }
         public int CompareTo(Isotope other)
         {
             return Index.CompareTo(other.Index);
